CRUD RESTful Suave-based capability data store
This example application shows how to build an F# Azure App Service
that implements a CRUD RESTful API on a capabilies-based data store.

It uses F# scripting to define the application, FAKE to build the
application as an executable for deployment, and FsCheck to do
properties-based testing that must pass before the Azure App Service
is deployed.